pyproject.toml
[project]
name = "nl2sql-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.115.14",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.5",
    "langchain-text-splitters>=0.3.8",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "uvicorn>=0.35.0",
]
